move limit_orders functions into own file UnsettledTradesByIDs and getSettledTrades both use cache now
