+++ server/routes/admin.router.js
@@ -59,80 +59,6 @@ router.get('/debug/:user_id', rejectUnauthenticated, async (req, res) => {
 router.post('/ordersReset', rejectUnauthenticated, async (req, res) => {
   res.sendStatus(500);
   return;
-  // look man, I probably never should have built this route in the first place
-  // but it was so useful for testing and I'm too lazy to delete it
-  if (req.user.admin) {
-    const queryText = `DROP TABLE IF EXISTS "limit_orders";
-    DROP TABLE IF EXISTS "orders";
-    CREATE TABLE IF NOT EXISTS "limit_orders"
-    (
-      order_id character varying COLLATE pg_catalog."default" NOT NULL,
-
-      "userID" integer,
-      original_buy_price numeric(32,16),
-      original_sell_price numeric(32,16),
-      trade_pair_ratio numeric(32,8),
-      flipped boolean DEFAULT false,
-      flipped_at timestamptz,
-      filled_at timestamptz,
-      reorder boolean DEFAULT false,
-      include_in_profit boolean DEFAULT true,
-      will_cancel boolean DEFAULT false,
-
-      product_id character varying COLLATE pg_catalog."default",
-      coinbase_user_id character varying COLLATE pg_catalog."default",
-      base_size numeric(32,8),
-      limit_price numeric(32,8),
-      post_only boolean,
-      side character varying COLLATE pg_catalog."default",
-      client_order_id character varying COLLATE pg_catalog."default",
-      next_client_order_id character varying COLLATE pg_catalog."default",
-      "status" character varying COLLATE pg_catalog."default",
-      time_in_force character varying COLLATE pg_catalog."default",
-      created_time timestamptz,
-      completion_percentage numeric(32,8),
-      filled_size numeric(32,8),
-      average_filled_price numeric(32,8),
-      fee numeric(32,8),
-      number_of_fills numeric(32,8),
-      filled_value numeric(32,8),
-      pending_cancel boolean,
-      size_in_quote boolean,
-      total_fees numeric(32,16),
-      previous_total_fees numeric(32,16),
-      size_inclusive_of_fees boolean,
-      total_value_after_fees numeric(32,16),
-      trigger_status character varying COLLATE pg_catalog."default",
-      order_type character varying COLLATE pg_catalog."default",
-      reject_reason character varying COLLATE pg_catalog."default",
-      settled boolean DEFAULT false,
-      product_type character varying COLLATE pg_catalog."default",
-      reject_message character varying COLLATE pg_catalog."default",
-      cancel_message character varying COLLATE pg_catalog."default",
-
-      CONSTRAINT orders_pkey PRIMARY KEY (order_id)
-    );
-    CREATE INDEX reorders
-    ON "limit_orders" ("side", "flipped", "will_cancel", "userID", "settled");`;
-    try {
-      await pool.query(queryText);
-      res.sendStatus(200);
-      // update orders on client for all users
-      userStorage.getAllUsers().forEach(userID => {
-        // update orders on client
-        messenger[userID].newMessage({
-          type: 'general',
-          text: `Orders table reset`,
-          orderUpdate: true
-        })
-      })
-    } catch (err) {
-      res.sendStatus(500);
-      devLog(err, 'error resetting orders table');
-    }
-  } else {
-    res.sendStatus(403)
-  }
 });
 
 /**
